Make Games a complete, self contained feature
