+++ src/features/Games/GamesList.js
@@ -1,29 +1,16 @@
 import React, { Component } from 'react';
-import {
-  Platform,
-  StyleSheet,
-  Text,
-  View,
-  Image,
-  Dimensions,
-  SectionList,
-  ScrollView,
-  TextInput,
-  ListView,
-  Button,
-  TouchableHighlight,
-  TouchableOpacity,
-  Alert,
-} from 'react-native';
+import { Text, View, Dimensions, SectionList, ScrollView, TextInput, TouchableHighlight, TouchableOpacity } from 'react-native';
 import ScalableImage from 'react-native-scalable-image';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import sectionListGetItemLayout from 'react-native-section-list-get-item-layout';
-import { homeButtonHeader } from '../src/utils/Headers';
-import { scale, verticalScale, moderateScale } from '../src/utils/Scaling';
+import { homeButtonHeader } from '../../utils/Headers';
+import { scale, verticalScale } from '../../utils/Scaling';
+import { styles } from './Games.styles';
+import AlphaBar from './AlphaBar';
 
 const ROW_HEIGHT = verticalScale(35);
 
-export default class GamesListScreen extends React.Component {
+export default class GamesList extends Component {
   static navigationOptions = ({ navigation, navigationOptions }) => {
     const { params } = navigation.state;
 
@@ -151,18 +138,16 @@ export default class GamesListScreen extends React.Component {
     let image = gameType.headerImageUrl;
 
     return (
-      <View style={{
-        flex: 1, backgroundColor: 'whitesmoke'
-      }}>
+      <View style={styles.listContainer}>
         <ScalableImage width={Dimensions.get('window').width}
           source={{uri: image}}/>
 
-        <View style={{height: verticalScale(34), margin: scale(8), borderRadius: scale(8), backgroundColor: 'white', flexDirection: 'row', justifyContent: 'flex-start', alignItems: 'center'}}>
-          <View style={{width: verticalScale(40), height: '100%', justifyContent: 'center', alignItems: 'center'}}>
+        <View style={styles.searchContainer}>
+          <View style={styles.searchView}>
             <Icon name={'search'} size={verticalScale(24)} color='lightgray' />
           </View>
           <TextInput
-            style={{flex: 1, height: '100%', padding: 0, textDecorationLine: 'none', fontSize: scale(20), fontFamily: 'Arial', color: 'gray'}}
+            style={styles.searchInput}
             placeholder="Search"
             placeholderTextColor='lightgray'
             onChangeText={this.handleChangeSearchText}
@@ -170,7 +155,7 @@ export default class GamesListScreen extends React.Component {
             value={this.state.searchFilter}
           />
           {this.state.searchFilter.length > 0 && 
-            <TouchableOpacity style={{width: verticalScale(40), height: '100%', justifyContent: 'center', alignItems: 'center'}} onPress={this.clearSearchText}>
+            <TouchableOpacity style={styles.searchClear} onPress={this.clearSearchText}>
               <Icon name={'remove'} size={verticalScale(24)} color='gray' />
             </TouchableOpacity>
           }
@@ -197,115 +182,3 @@ export default class GamesListScreen extends React.Component {
   }
 }
 
-class AlphaBar extends React.Component{
-  constructor(props) {
-    super(props);
-
-    alphabet = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
-    this.buttons = alphabet.split('');
-    this.buttons.unshift('0-9');
-
-    this.buttonLayout = {};
-  }
-
-  alphaScroll: ?ScrollView;
-
-  setScroll = (el) => {
-    this.alphaScroll = el;
-  }
-
-  layoutScroll = (e) => {
-    this.scrollWidth = e.nativeEvent.layout.width;
-  }
-
-  handleContentSizeChange = (contentHeight, contentWidth) => {
-    this.scrollContentWidth = contentWidth;
-  };
-
-  layoutButton = (e, key) => {
-    if (!this.buttonLayout[key]) {
-      this.buttonLayout[key] = {width: e.nativeEvent.layout.width, x: e.nativeEvent.layout.x};
-    }
-  }
-
-  scrollToButton = (key, animate = true) => {
-    const scrollHalfWidth = this.scrollWidth * 0.5;
-    const button = this.buttonLayout[key];
-    const buttonCenter = button.x + (button.width * 0.5);
-
-    let scrollPos = buttonCenter - scrollHalfWidth;
-
-    scrollPos = scrollPos < 0 ? 0 : scrollPos;
-
-    this.alphaScroll.scrollTo({x: scrollPos, animated: animate});
-  }
-
-  handlePress = (key) => {
-    this.scrollToButton(key, true);
-    this.props.onPress(key);
-  }
-
-  render() {
-    return (
-      <ScrollView style={{backgroundColor: '#555555'}} contentContainerStyle={{backgroundColor: '#555555'}} showsHorizontalScrollIndicator={false} horizontal={true} ref={this.setScroll} onLayout={this.layoutScroll} onContentSizeChange={this.handleContentSizeChange}>
-        <View style={{flexDirection: 'row'}}>
-          {this.buttons.map(text => {
-            let style = [styles.letter];
-            if (text.length > 1) {
-              style.push(styles.wideLetter);
-            }
-            return (
-              <TouchableHighlight style={style}
-                key={text}
-                onLayout={(e) => {this.layoutButton(e, text)}}
-                onPress={() => this.handlePress(text)}>
-                  <Text style={styles.scrollLetterText}>{text}</Text>
-              </TouchableHighlight>
-            );
-          })}
-        </View>
-      </ScrollView>
-    );
-  }
-}
- 
-const styles=StyleSheet.create({
-  stretch: {
-   flex: 1,
-   width: '100%',
-  },
-  letter: {
-    backgroundColor: '#555555',
-    alignItems: 'center',
-    borderLeftWidth: StyleSheet.hairlineWidth * 3,
-    borderLeftColor: 'black',
-    width: scale(35),
-    margin: 0,
-  },
-  wideLetter: {
-    width: scale(48)
-  },
-  scrollLetterText: {
-    fontSize: scale(24),
-    fontFamily: 'arial',
-    textAlign: 'center',
-    color: '#e3e4e4',
-    paddingBottom: verticalScale(3),
-    paddingTop: verticalScale(3),
-  },
-  item1: {
-    height: ROW_HEIGHT,
-    paddingHorizontal: scale(5),
-    borderBottomWidth: StyleSheet.hairlineWidth * 2,
-    borderColor: '#9ca4ab',
-    justifyContent: 'center',
-  },
-  item2: {
-    backgroundColor: '#dddddd',
-  },
-  itemText: {
-    fontSize: scale(18),
-    fontFamily: 'Arial',
-    paddingRight: scale(12)
-  }
-});
